feat(rusmppc/client): added auto_enquire_link_response. See 41d2139
When enabled, the connection will automatically respond to any enquire link requests received from the server.
When disabled, the client will need to handle enquire link requests manually. The enquire link command will be received as an event in the event stream.

Enabled by default, which has the same behavior as before adding this feature

Signed-off-by: Jad K. Haddad <jadkhaddad@gmail.com>